Move sloppy tests into their own file.
