+++ .gitignore
@@ -4,8 +4,7 @@ node_modules
 test/mocha.js
 test/mocha.css
 
-test/tests.es5.js
-test/async.es5.js
+test/*.es5.js
 test/tests.browser.js
 
 .idea
